Reuse same map for public and private map in explore page
