+++ app/views/ExploreDeep/MapView/styles.css
@@ -2,32 +2,4 @@
     display: flex;
     position: relative;
     z-index: 0;
-
-    .map {
-        width: 100%;
-        height: 80vh;
-    }
-
-    :global {
-        .mapboxgl-popup {
-            width: 400px;
-            max-width: 400px !important;
-            height: calc(40vh - 100px);
-
-            .mapboxgl-popup-content {
-                overflow-y: auto !important;
-            }
-        }
-    }
-}
-
-.count {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    outline: 0;
-    border: none;
-    border-radius: 50%;
-    background-color: #1a3ed0;
-    color: #fff;
 }
